Install uuid
Added the uuid package to the project dependencies and updated the package.json file. This package is essential for generating unique identifiers (UUIDs) that can be used for various purposes, such as assigning unique IDs to waifus or any other entities in the application. By installing uuid, we have the necessary functionality to generate unique identifiers in a reliable and standardized manner