Implemented POST and GET routes for waifus endpoint
Added the POST route to handle requests for adding a new waifu to the waifus array. The waifu object received from the request body is assigned a unique ID generated using the uuid module. The waifu is then pushed into the waifus array. Also implemented the GET route to retrieve a waifu based on its ID. The route searches for the waifu with the matching ID in the waifus array and sends it as the response