Implemented DELETE method for waifus endpoint
Added the DELETE method to the waifus endpoint in order to handle requests for deleting a waifu based on the provided ID. When a request with the specified ID is received, the corresponding waifu is removed from the endpoint's data source. The response confirms the successful deletion of the waifu from the endpoint.