Updated waifu.json with new data
Uploaded an updated version of the waifu.json file to reflect the addition of new data. After performing a POST request to add a new waifu, the waifu.json file now contains the updated list of waifus, including the newly added data. This allows the file to accurately represent the current state of the waifus.